refactor: organize legend and its overrides
