+++ packages/superset-ui-preset-chart-xy/src/BoxPlot/transformProps.ts
@@ -1,26 +1,9 @@
+/* eslint-disable sort-keys, no-magic-numbers */
+
+import { pick } from 'lodash';
 import { ChartProps } from '@superset-ui/chart';
 import { BoxPlotDataRow, RawBoxPlotDataRow } from './types';
-
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-/* eslint-disable sort-keys, no-magic-numbers */
+import { HookProps } from './BoxPlot';
 
 export default function transformProps(chartProps: ChartProps) {
   const { width, height, formData, payload } = chartProps;
@@ -57,6 +40,17 @@ export default function transformProps(chartProps: ChartProps) {
     encoding.y.scale.domain = valueDomain;
   }
 
+  const hooks = chartProps.hooks as HookProps;
+
+  const fieldsFromHooks: (keyof HookProps)[] = [
+    'TooltipRenderer',
+    'LegendRenderer',
+    'LegendGroupRenderer',
+    'LegendItemRenderer',
+    'LegendItemMarkRenderer',
+    'LegendItemLabelRenderer',
+  ];
+
   return {
     data,
     width,
@@ -64,5 +58,6 @@ export default function transformProps(chartProps: ChartProps) {
     margin,
     theme,
     encoding,
+    ...pick(hooks, fieldsFromHooks),
   };
 }
