+++ packages/superset-ui-preset-chart-xy/src/ScatterPlot/transformProps.ts
@@ -1,11 +1,23 @@
-import { ChartProps } from '@superset-ui/chart';
-
 /* eslint-disable sort-keys */
+
+import { pick } from 'lodash';
+import { ChartProps } from '@superset-ui/chart';
+import { HookProps } from './ScatterPlot';
 
 export default function transformProps(chartProps: ChartProps) {
   const { width, height, formData, payload } = chartProps;
   const { encoding, margin, theme } = formData;
   const { data } = payload;
+  const hooks = chartProps.hooks as HookProps;
+
+  const fieldsFromHooks: (keyof HookProps)[] = [
+    'TooltipRenderer',
+    'LegendRenderer',
+    'LegendGroupRenderer',
+    'LegendItemRenderer',
+    'LegendItemMarkRenderer',
+    'LegendItemLabelRenderer',
+  ];
 
   return {
     data,
@@ -14,5 +26,6 @@ export default function transformProps(chartProps: ChartProps) {
     encoding,
     margin,
     theme,
+    ...pick(hooks, fieldsFromHooks),
   };
 }
